demo/src/test/java/com/springreactshop/demo/service/MemberServiceTest.java
package com.springreactshop.demo.service;

import com.springreactshop.demo.domain.Member;
import com.springreactshop.demo.repository.MemberRepository;
import com.springreactshop.demo.representation.JwtRequest;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.transaction.annotation.Transactional;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
class MemberServiceTest {

    @Autowired
    MemberService memberService;

    @Autowired
    MemberRepository memberRepository;


    @Test
    public void 회원추가() {
        //Given
        JwtRequest signupRequest = new JwtRequest();
        signupRequest.setUsername("testUser");
        signupRequest.setPassword("1234");

        //When
        String username = memberService.signUpUser(signupRequest);

        //Then
        assertThat(username).isEqualTo("testUser");
    }
}